Samplerates should be floats, not ints

+++ psychic/utils.py
@@ -57,7 +57,7 @@ def spectrogram(signal, nfft, stepsize):
 
 def get_samplerate(d):
   '''Derive the sample rate from the timestamps d.I[0]'''
-  return int(np.round(1./np.median(np.diff(d.I[0]))))
+  return np.round(1./np.median(np.diff(d.I[0])))
 
 def slice(d, markers_to_class, offsets):
   '''
